# app/backend/contracts/openapi.snapshot.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/vision/governance/shadow": {
      "get": {
        "operationId": "GovernanceController_shadow",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Is shadow ingestion configured and landing?",
        "tags": [
          "vision-governance"
        ]
      }
    },
    "/api/vision/governance/comparison": {
      "get": {
        "operationId": "GovernanceController_comparison",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The paired comparison: same scans, same ground truth, both providers.",
        "tags": [
          "vision-governance"
        ]
      }
    },
    "/api/vision/governance/drift": {
      "get": {
        "operationId": "GovernanceController_drift",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Is the incumbent still the provider we promoted?",
        "tags": [
          "vision-governance"
        ]
      }
    },
    "/api/vision/governance/recommendation": {
      "get": {
        "operationId": "GovernanceController_recommendation",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "PROMOTE / MAINTAIN / DEMOTE / HOLD / REQUIRE_MORE_DATA — with its evidence.",
        "tags": [
          "vision-governance"
        ]
      }
    },
...
    "/api/vision/canary": {
      "get": {
        "operationId": "CanaryController_plan",
        "parameters": [
          {
            "name": "atPercent",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The complete, deterministic canary progression plan.",
        "tags": [
          "vision-canary"
        ]
      }
    },
    "/api/vision/canary/readiness": {
      "get": {
        "operationId": "CanaryController_readiness",
        "parameters": [
          {
            "name": "atPercent",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The next-move recommendation with its three explained signals.",
        "tags": [
          "vision-canary"
        ]
      }
    },
    "/api/vision/canary/stages": {
      "get": {
        "operationId": "CanaryController_stages",
        "parameters": [
          {
            "name": "atPercent",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current and next rung, exposure, and what must hold to advance.",
        "tags": [
          "vision-canary"
        ]
      }
    },
    "/api/vision/canary/checklist": {
      "get": {
        "operationId": "CanaryController_checklist",
        "parameters": [
          {
            "name": "atPercent",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Monitoring and verification checklists.",
        "tags": [
          "vision-canary"
        ]
      }
    },
    "/api/vision/canary/timeline": {
      "get": {
        "operationId": "CanaryController_timeline",
        "parameters": [
          {
            "name": "atPercent",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The full positioned ladder.",
        "tags": [
          "vision-canary"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vitals Fit API",
    "description": "Nutrition platform API. Endpoints under /vision/{rollout,governance,promotion,rollback,canary} and the platform-wide /vision/learning analytics are OPERATOR-only (ADMIN_EMAILS allowlist).",
    "version": "5.5",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "InternalErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 500
          },
          "message": {
            "type": "string",
            "example": "Internal server error",
            "description": "Always this literal string — never the real cause."
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-22T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/vision/scans",
            "description": "The route that failed, for log correlation."
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
      "HttpErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "Repeats the HTTP status code."
          },
          "message": {
            "description": "A human-readable message, or one string per failed validation rule.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Resource not found."
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Nest's standard reason phrase for the status code."
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
    }
  }
}