# app/backend/src/vision/canary/canary.controller.ts
import { Controller, Get, Query, UseGuards } from '@nestjs/common';
import { ApiBearerAuth, ApiQuery, ApiTags } from '@nestjs/swagger';
import { JwtAuthGuard } from '../../auth/guards/jwt-auth.guard';
import { AdminGuard } from '../../common/guards/admin.guard';
import { parseDays, parsePercent } from '../../common/http/query-parsers';
import { CanaryEngine } from './canary.engine';
import { ApiAdminErrors } from '../../common/swagger/error-responses';

const DAYS_QUERY = { name: 'days', required: false, type: String, description: 'Lookback window in days (1–3650).' };
const AT_PERCENT_QUERY = {
  name: 'atPercent',
  required: false,
  type: String,
  description:
    "The operator's stated current rollout position, 0–100 (default 0). The platform persists no live canary state.",
};

/**
 * Canary progression API (V4.4) — GET only, read-only by construction. Every
 * route derives a recommendation (or a slice of one); none can advance traffic,
 * change a flag, or write anything. The plan is a document a human reads and
 * acts on.
 *
 * `atPercent` is the operator's stated current position (0..100; default 0 =
 * pre-rollout), since the platform persists no live canary state.
 */
@ApiTags('vision-canary')
@ApiBearerAuth()
@ApiAdminErrors()
@ApiQuery(AT_PERCENT_QUERY)
@ApiQuery(DAYS_QUERY)
@UseGuards(JwtAuthGuard, AdminGuard)
@Controller('vision')
export class CanaryController {
  constructor(private readonly engine: CanaryEngine) {}

  /** The complete, deterministic canary progression plan. */
  @Get('canary')
  plan(@Query('atPercent') atPercent?: string, @Query('days') days?: string) {
    return this.engine.plan(parsePercent(atPercent), parseDays(days));
  }

  /** The next-move recommendation with its three explained signals. */
  @Get('canary/readiness')
  readiness(@Query('atPercent') atPercent?: string, @Query('days') days?: string) {
    return this.engine.readiness(parsePercent(atPercent), parseDays(days));
  }

  /** Current and next rung, exposure, and what must hold to advance. */
  @Get('canary/stages')
  stages(@Query('atPercent') atPercent?: string, @Query('days') days?: string) {
    return this.engine.stages(parsePercent(atPercent), parseDays(days));
  }

  /** Monitoring and verification checklists. */
  @Get('canary/checklist')
  checklist(@Query('atPercent') atPercent?: string, @Query('days') days?: string) {
    return this.engine.checklist(parsePercent(atPercent), parseDays(days));
  }

  /** The full positioned ladder. */
  @Get('canary/timeline')
  timeline(@Query('atPercent') atPercent?: string, @Query('days') days?: string) {
    return this.engine.timeline(parsePercent(atPercent), parseDays(days));
  }
}